pyproject.toml
[project]
name = "vegapunk-agent"
version = "0.1.0"
description = "Research & Teaching Agent - LangGraph-based educational content generator"
authors = [
    {name = "Your Name", email = "your@email.com"}
]
readme = "README.md"
requires-python = ">=3.10"
dependencies = [
    "langgraph>=0.2.0",
    "langchain-core>=0.3.0",
    "langchain-openai>=0.2.0",
    "langchain-anthropic>=0.2.0",
    "openai>=1.12.0",
    "anthropic>=0.18.0",
    "tavily-python>=0.5.0",
    "python-dotenv>=1.0.0",
    "GitPython>=3.1.40",
]

[project.scripts]
vegapunk = "main:main"

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.hatch.build.targets.wheel]
packages = ["src"]

[dependency-groups]
dev = []
